requirements.txt
numpy==2.1.1
openai==1.47.1
PyPDF2==3.0.1
python-dotenv==1.0.1
scikit_learn==1.5.2
setuptools==70.3.0
slack_sdk==3.33.1
